roo-code/gemini-2.5-pro/initial/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases[mysql]
mysqlclient
python-jose[cryptography] # JWTトークン用
passlib[bcrypt] # パスワードハッシュ化用
python-multipart # ファイルアップロード用 (将来的に必要になる可能性)